requirements.txt
# Flask Framework
flask==3.0.3
flask-cors==4.0.0

# Configuration Management
python-dotenv==1.0.1

# AI/ML Libraries
openai==1.40.2
nltk==3.8.1
textblob==0.17.1
numpy==1.24.3
scikit-learn==1.3.0

# Data Validation
pydantic==2.8.2

# HTTP Requests
requests==2.32.3

# Async Support
aiohttp==3.9.5

# Caching
cachetools==5.3.3

# Rate Limiting
flask-limiter==3.8.0

# Monitoring and Metrics
prometheus-client==0.20.0

# Testing
pytest==8.2.2
pytest-cov==5.0.0
pytest-asyncio==0.23.7
pytest-mock==3.14.0

# Code Quality
black==24.4.2
flake8==7.1.0
isort==5.13.2
mypy==1.10.1

# Security
werkzeug==3.0.3

# Utilities
click==8.1.7
python-dateutil==2.9.0

# Development
pre-commit==3.7.1
